Moved swap logic to PuzzleState, removed getPuzzle()
